requirements.txt
flask==3.0.2
flask-cors==4.0.0
supabase==2.3.4
python-dotenv==1.0.1
gunicorn==21.2.0
pandas==2.2.1